requirements.txt
pandas 
numpy
yfinance
scikit-learn
gspread
oauth2client
pyTelegramBotAPI
plotly yfinance